Added security to backend with authorization

+++ dashboard-backend/src/main/java/com/weissdennis/leeachaanbot/api/v1/CustomCommandController.java
@@ -3,11 +3,11 @@ package com.weissdennis.leeachaanbot.api.v1;
 import com.weissdennis.leeachaanbot.model.CustomCommand;
 import com.weissdennis.leeachaanbot.persistence.Customcommands;
 import com.weissdennis.leeachaanbot.service.CustomCommandService;
+import com.weissdennis.leeachaanbot.service.PermissionsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Repository;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -18,10 +18,12 @@ import java.util.List;
 public class CustomCommandController {
 
     private final CustomCommandService customCommandService;
+    private final PermissionsService permissionsService;
 
     @Autowired
-    public CustomCommandController(CustomCommandService customCommandService) {
+    public CustomCommandController(CustomCommandService customCommandService, PermissionsService permissionsService) {
         this.customCommandService = customCommandService;
+        this.permissionsService = permissionsService;
     }
 
     @RequestMapping(path = "/all", method = RequestMethod.GET)
@@ -39,22 +41,34 @@ public class CustomCommandController {
     }
 
     @RequestMapping(path = "/add", method = RequestMethod.POST)
-    public HttpEntity<Customcommands> addCustomCommand(@RequestBody CustomCommand customCommand) {
-        return new ResponseEntity<>(customCommandService.addCustomCommand(customCommand), HttpStatus.OK);
+    public HttpEntity<Customcommands> addCustomCommand(@RequestBody CustomCommand customCommand,
+                                                       @RequestHeader("Authorization") String authorization) {
+        if (permissionsService.hasAdministrationRights(authorization)) {
+            return new ResponseEntity<>(customCommandService.addCustomCommand(customCommand), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @RequestMapping(path = "/update", method = RequestMethod.POST)
-    public HttpEntity<Customcommands> updateCustomCommand(@RequestBody Customcommands customCommand) {
-        Customcommands updatedCustomCommand = customCommandService.updateCustomCommand(customCommand);
-        if (updatedCustomCommand != null) {
-            return new ResponseEntity<>(updatedCustomCommand, HttpStatus.OK);
+    public HttpEntity<Customcommands> updateCustomCommand(@RequestBody Customcommands customCommand,
+                                                          @RequestHeader("Authorization") String authorization) {
+        if (permissionsService.hasAdministrationRights(authorization)) {
+            Customcommands updatedCustomCommand = customCommandService.updateCustomCommand(customCommand);
+            if (updatedCustomCommand != null) {
+                return new ResponseEntity<>(updatedCustomCommand, HttpStatus.OK);
+            }
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @RequestMapping(path = "/delete/{id}", method = RequestMethod.DELETE)
-    public HttpEntity<Void> deleteCustomCommand(@PathVariable String id) {
-        boolean deleted = customCommandService.deleteCustomCommand(id);
-        return new ResponseEntity<>(deleted ? HttpStatus.OK : HttpStatus.NOT_FOUND);
+    public HttpEntity<Void> deleteCustomCommand(@PathVariable String id,
+                                                @RequestHeader("Authorization") String authorization) {
+        if (permissionsService.hasAdministrationRights(authorization)) {
+            boolean deleted = customCommandService.deleteCustomCommand(id);
+            return new ResponseEntity<>(deleted ? HttpStatus.OK : HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 }
